Update README.md
These wines are especially prized for their freshness and quality.
The code in this project runs a scatter graph that displays ratings of professional wine tasters. Each rating is represented by a colored dot on the graph.
The dots are arranged on an xy plot, with the x-axis representing pH (lower values are more acidic; higher values are more basic) and with the y-axis representing percentage of alcohol content.

What the scatter plot appears to show is that vinho verde wines with a higher percentage of acohol content are generally given higher taster ratings than those with a lower percentage of alcohol content.